first_update

+++ accreapp/views.py
@@ -16,6 +16,7 @@ from django.contrib.contenttypes.models import ContentType
 
 from accreapp.models import TaggedWhatever, Profile, File, Category
 from taggit.models import Tag
+from actstream import action
 
 import os, json
 from model_utils import Choices
@@ -24,7 +25,7 @@ from model_utils import Choices
 def home(request):
     current_user = request.user
     user_instance = User.objects.get(id=current_user.id)
-    print (user_instance.get_group_permissions())
+    #print (user_instance.get_group_permissions())
     return render(request, 'accreapp/home.html')
 
 def index(request):
@@ -48,20 +49,23 @@ def logout_view(request):
     return render(request, 'accreapp/index.html')
 
 def callUpload(request):
-    if request.user.is_authenticated and request.is_ajax() and request.method == 'POST':
+    if request.user.is_authenticated and request.method == 'POST':
         blob = request.FILES.get('image')
         fileN = request.POST.get('fname')
         orderN = request.POST.get('order')
         current_user = request.user
+        content_type = ContentType.objects.get(id=9)
         user_instance = User.objects.get(id=current_user.id)
         if not os.path.exists(default_storage.path(fileN)):
             b = File(user=user_instance,file_name=fileN,description='To be updated...',extracted_text='To be updated...' )
-            b.save()
+            action.send(user_instance, verb='Added',action_object=b,description='Added File  '+fileN, target=content_type)
             default_storage.save(fileN, blob)
+            b.save()
             return JsonResponse({'is_uploaded':1,'filename': fileN, 'order':int(orderN) })
         else:
             #print 'Already saved on the server -->'+fileN
-            return JsonResponse({'is_uploaded':0})
+            #return JsonResponse({'is_uploaded':0})
+            raise Http404
     else:
         raise Http404
 
@@ -71,6 +75,7 @@ ORDER_COLUMN_CHOICES =Choices(
     ('3', 'description'),
     ('4', 'tagged_items__tag__code'),
     ('5', 'uploaded_at'),
+    ('6', 'id'),
 )
 
 def mainTableData(request):
@@ -100,7 +105,7 @@ def mainTableData(request):
         if search_value:
             allfiles = allfiles.filter(Q(file_name__icontains=search_value) |
                                     Q(tags__code__icontains=search_value) |
-                                    Q(description=search_value)).all()
+                                    Q(description__icontains=search_value)).all().distinct()
         count = allfiles.count()
         if length != -1:
             allfiles = allfiles.order_by(order_column)[start:start + length]
@@ -109,11 +114,14 @@ def mainTableData(request):
         
         lst = []
         for i in allfiles:
+            dd = i.date_created
+            if i.date_created is not None:
+                dd = i.date_created.strftime("%B %d, %Y")
             lst.append({
                 'id':i.id,
                 'file_name': str(i.file_name),
                 'tagged_items__tag__code': i.get_tags(),
-                'date_created': i.date_created,
+                'date_created': dd,
                 'description': i.description,
                 'uploaded_at': i.uploaded_at
             })
@@ -131,37 +139,47 @@ def mainTableData(request):
 
 def bulkUpdate(request):
     if request.user.is_authenticated and request.method == 'POST':
+        current_user = request.user
+        user_instance = User.objects.get(id=current_user.id)
         controlNums = json.loads(request.POST['controlNums'])
         description = request.POST.get('description')
         date_created = request.POST.get('date_created')
-        tags =  request.POST.get('tags').split(',')
-        
+        tags =  request.POST.getlist('tags')
+        print (tags)
         is_updated = 1
-        #print (tags)
-        #print (len(tags))
         d = {}
+        content_type = ContentType.objects.get(id=9)
+        
         if len(controlNums) > 0 :
             if not description == '':
                 d['description'] = description
             if not date_created == '':
                 d['date_created'] = date_created
-            if len(tags) > 0 and tags[0] != '':
-                bulk_taggedFile = []
-                content_type = ContentType.objects.get(id=9)
-                #print (content_type)
-                for i in controlNums:
-                    #print ('iiiiiiii--> '+str(i))
-                    for j in tags:
-                        #print ('j--> '+str(j))
+            bulk_taggedFile = []
+            for i in controlNums:
+                b = File.objects.get(id=i)
+                codes = []
+                for j in tags:
+                    if len(tags) > 0 and tags[0] != '':
                         if Category.objects.filter(id=j).exists():
                             tagss =  Category.objects.get(id = j)
                             TaggedWhatever.objects.filter(object_id=i).delete()
                             bulk_taggedFile.append(TaggedWhatever(object_id = i,content_type = content_type,tag = tagss))
+                            TaggedWhatever.objects.bulk_create(bulk_taggedFile)
+                            codes.append(tagss.code)      
                         else:
                             print ('invalid tag')
-                #print (bulk_taggedFile)
-                TaggedWhatever.objects.bulk_create(bulk_taggedFile)
+                if len(tags) > 0 and tags[0] != '':
+                    action.send(user_instance, verb='Tagged',action_object=b,description='Tagged File  '+str(b.file_name)+' on '+','.join(codes) , target=content_type)
+                if not description == '' and not date_created == '':   
+                    action.send(user_instance, verb='Updated',action_object=b,description='Updated File  '+str(b.file_name)+'\'s description and date.' , target=content_type)
+                elif not description == '':    
+                    print('i am here')    
+                    action.send(user_instance, verb='Updated',action_object=b,description='Updated File  '+str(b.file_name)+'\'s description.' , target=content_type)
+                elif not date_created == '':     
+                    action.send(user_instance, verb='Updated',action_object=b,description='Updated File  '+str(b.file_name)+'\'s date.' , target=content_type)
             File.objects.filter(id__in=controlNums).update(**d)
+           
         return JsonResponse({'is_updated': is_updated})
     else:
         raise Http404
@@ -169,12 +187,22 @@ def bulkUpdate(request):
 def bulkDelete(request):
     current_user = request.user
     user_instance = User.objects.get(id=current_user.id)
-    print (user_instance.get_group_permissions())
+    content_type = ContentType.objects.get(id=9)
+
     if request.user.is_authenticated and user_instance.has_perm('accreapp.delete_file'):
         controlNums = json.loads(request.POST['controlNums'])
         is_deleted = 1
+        counter = 0
         if len(controlNums) > 0 :
-            File.objects.filter(id__in=controlNums).delete()
+            for i in controlNums:
+                counter+=1
+                b = File.objects.get(id=i)
+                print (b)
+                print (str(counter))
+                action.send(user_instance, verb='Deleted',action_object=b,description='Deleted File  '+str(b.file_name) , target=content_type)
+            if counter == len(controlNums):
+                print ('wwwww')
+                File.objects.filter(id__in=controlNums).delete()
             return JsonResponse({'is_deleted_items': is_deleted})
         else:
             return JsonResponse({'is_deleted_items': 0})
